requirements.txt
mysql-connector-python
requests
schedule
pywhatkit
python-dotenv
